autoClose feature and minute-draggability incompatibility.
When "autoClose=true", the minute-draggability cause the minutes to be
marked as "undefined" when the "currently selected minute" don't match
one of the minute on the dial.  Also, as we need to click the time to
close the dial, clicking the already-selected minute don't do anything
(as this is the drag-fake-needle).  To avoid that, disable the
draggability feature on minute dial when autoClose=true.